Struct embassy_time::Instant

pub struct Instant { /* private fields */ }

An Instant in time, based on the MCU’s clock ticks since startup.

Implementations

impl Instant

pub const MIN: Instant = _

The smallest (earliest) value that can be represented by the Instant type.

pub const MAX: Instant = _

The largest (latest) value that can be represented by the Instant type.

pub fn now() -> Instant

Returns an Instant representing the current time.

pub const fn from_ticks(ticks: u64) -> Self

Create an Instant from a tick count since system boot.

pub const fn from_micros(micros: u64) -> Self

Create an Instant from a microsecond count since system boot.

pub const fn from_millis(millis: u64) -> Self

Create an Instant from a millisecond count since system boot.

pub const fn from_secs(seconds: u64) -> Self

Create an Instant from a second count since system boot.

pub const fn as_ticks(&self) -> u64

Tick count since system boot.

pub const fn as_secs(&self) -> u64

Seconds since system boot.

pub const fn as_millis(&self) -> u64

Milliseconds since system boot.

pub const fn as_micros(&self) -> u64

Microseconds since system boot.

pub fn duration_since(&self, earlier: Instant) -> Duration

Duration between this Instant and another Instant Panics on over/underflow.

pub fn checked_duration_since(&self, earlier: Instant) -> Option<Duration>

Duration between this Instant and another Instant

pub fn saturating_duration_since(&self, earlier: Instant) -> Duration

Returns the duration since the “earlier” Instant. If the “earlier” instant is in the future, the duration is set to zero.

pub fn elapsed(&self) -> Duration

Duration elapsed since this Instant.

pub fn checked_add(&self, duration: Duration) -> Option<Instant>

Adds one Duration to self, returning a new Instant or None in the event of an overflow.

pub fn checked_sub(&self, duration: Duration) -> Option<Instant>

Subtracts one Duration to self, returning a new Instant or None in the event of an overflow.