embassy_time
Struct Duration

pub struct Duration { /* private fields */ }

Represents the difference between two Instants

Implementations

impl Duration

pub const MIN: Duration

The smallest value that can be represented by the Duration type.

pub const MAX: Duration

The largest value that can be represented by the Duration type.

pub const fn as_ticks(&self) -> u64

Tick count of the Duration.

pub const fn as_secs(&self) -> u64

Convert the Duration to seconds, rounding down.

pub const fn as_millis(&self) -> u64

Convert the Duration to milliseconds, rounding down.

pub const fn as_micros(&self) -> u64

Convert the Duration to microseconds, rounding down.

pub const fn from_ticks(ticks: u64) -> Duration

Creates a duration from the specified number of clock ticks

pub const fn from_secs(secs: u64) -> Duration

Creates a duration from the specified number of seconds, rounding up.

pub const fn from_millis(millis: u64) -> Duration

Creates a duration from the specified number of milliseconds, rounding up.

pub const fn from_micros(micros: u64) -> Duration

Creates a duration from the specified number of microseconds, rounding up. NOTE: Delays this small may be inaccurate.

pub const fn from_nanos(micros: u64) -> Duration

Creates a duration from the specified number of nanoseconds, rounding up. NOTE: Delays this small may be inaccurate.

pub const fn from_secs_floor(secs: u64) -> Duration

Creates a duration from the specified number of seconds, rounding down.

pub const fn from_millis_floor(millis: u64) -> Duration

Creates a duration from the specified number of milliseconds, rounding down.

pub const fn from_micros_floor(micros: u64) -> Duration

Creates a duration from the specified number of microseconds, rounding down. NOTE: Delays this small may be inaccurate.

pub const fn from_hz(hz: u64) -> Duration

Creates a duration corresponding to the specified Hz. NOTE: Giving this function a hz >= the TICK_HZ of your platform will clamp the Duration to 1 tick. Doing so will not deadlock, but will certainly not produce the desired output.

pub fn checked_add(self, rhs: Duration) -> Option<Duration>

Adds one Duration to another, returning a new Duration or None in the event of an overflow.

pub fn checked_sub(self, rhs: Duration) -> Option<Duration>

Subtracts one Duration to another, returning a new Duration or None in the event of an overflow.

pub fn checked_mul(self, rhs: u32) -> Option<Duration>

Multiplies one Duration by a scalar u32, returning a new Duration or None in the event of an overflow.

pub fn checked_div(self, rhs: u32) -> Option<Duration>

Divides one Duration a scalar u32, returning a new Duration or None in the event of an overflow.